--- src/routes/about.svelte
@@ -1,5 +1,0 @@
-<script>
-	import About from "../pages/About.svelte";
-</script>
-
-<About />
