--- src/routes/events/awards-for-theresia-stich.svelte
@@ -1,5 +1,0 @@
-<script>
-  import AwardsForTH from "../../pages/events/AwardsForTH.svelte";
-</script>
-
-<AwardsForTH />
