+++ src/routes/events/awards-for-theresia-stich/+page.svelte
@@ -1,10 +1,4 @@
 <script>
-  import { onMount } from "svelte";
-  import { routeStore as route_to } from "../../store/route.js";
-  import { events } from "../../data/events.js";
-  import { get_id } from "../../js/utils.js";
-
-  onMount(() => ($route_to = get_id(events, 5).route_to));
 </script>
 
 <div class="is-justify pb-4">
