--- src/routes/events/biomaterials-and-regenerative-cells.svelte
@@ -1,5 +1,0 @@
-<script>
-  import BiomaterialsAndRegenerativeCells from "../../pages/events/BiomaterialsAndRegenerativeCells.svelte";
-</script>
-
-<BiomaterialsAndRegenerativeCells />
