--- src/routes/events/btm-implant-bone-interface.svelte
@@ -1,5 +1,0 @@
-<script>
-  import BtmImplantBoneInterface from "../../pages/events/BtmImplantBoneInterface.svelte";
-</script>
-
-<BtmImplantBoneInterface />
