--- src/routes/events/first-workshop.svelte
@@ -1,5 +1,0 @@
-<script>
-  import FirstWorkshop from "../../pages/events/FirstWorkshop.svelte";
-</script>
-
-<FirstWorkshop />
