--- src/routes/events/mcp-biomimetic-calcium-deposition.svelte
@@ -1,5 +1,0 @@
-<script>
-  import McpBiomimeticCalciumDeposition from "../../pages/events/McpBiomimeticCalciumDeposition.svelte";
-</script>
-
-<McpBiomimeticCalciumDeposition />
