--- src/routes/events/online-seminar.svelte
@@ -1,5 +1,0 @@
-<script>
-  import OnlineSeminar from "../../pages/events/OnlineSeminar.svelte";
-</script>
-
-<OnlineSeminar />
