--- src/routes/events/osteomet-seminar-and-workshop.svelte
@@ -1,5 +1,0 @@
-<script>
-  import OsteometSeminarAndWorkshop from "../../pages/events/OsteometSeminarAndWorkshop.svelte";
-</script>
-
-<OsteometSeminarAndWorkshop />
