+++ src/routes/events/osteomet-seminar-and-workshop/+page.svelte
@@ -1,11 +1,5 @@
 <script>
-  import Icon from "../../components/Icon.svelte";
-  import { onMount } from "svelte";
-  import { routeStore as route_to } from "../../store/route.js";
-  import { events } from "../../data/events.js";
-  import { get_id } from "../../js/utils.js";
-
-  onMount(() => ($route_to = get_id(events, 2).route_to));
+  import Icon from "../../../components/Icon.svelte";
 </script>
 
 <h2 class="article-title">
