--- src/routes/events/rsc-advances.svelte
@@ -1,5 +1,0 @@
-<script>
-  import RscAdvances from "../../pages/events/RscAdvances.svelte";
-</script>
-
-<RscAdvances />
