--- src/routes/events/sai-porous-micro-structured-titanium.svelte
@@ -1,5 +1,0 @@
-<script>
-  import SaiPorousMicroStructuredTitanium from "../../pages/events/SaiPorousMicroStructuredTitanium.svelte";
-</script>
-
-<SaiPorousMicroStructuredTitanium />
