--- src/routes/events/second-workshop.svelte
@@ -1,5 +1,0 @@
-<script>
-  import SecondWorkshop from "../../pages/events/SecondWorkshop.svelte";
-</script>
-
-<SecondWorkshop />
