--- src/routes/events/surface-science.svelte
@@ -1,5 +1,0 @@
-<script>
-  import SurfaceScience from "../../pages/events/SurfaceScience.svelte";
-</script>
-
-<SurfaceScience />
