--- src/routes/gallery.svelte
@@ -1,5 +1,0 @@
-<script>
-  import Gallery from "../pages/Gallery.svelte";
-</script>
-
-<Gallery />
