--- src/routes/group.svelte
@@ -1,5 +1,0 @@
-<script>
-  import Group from "../pages/Group.svelte";
-</script>
-
-<Group />
