--- src/store/route.js
@@ -1,3 +1,0 @@
-import { writable } from "svelte/store";
-
-export const routeStore = writable("/");
